Move IDesignTimeDbContextFactory to DatabaseFactory
